# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Lotofond API'
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /trades:
    put:
      parameters:
        - name: Content-Language
          in: header
          description: Заголовок для определения локали (ru, en, zh_CN)
          required: true
        - in: query
          name: page
          schema:
            type: integer
          required: true
          description: Номер страницы для серверной пагинации
      tags:
        - Trades
      summary: Получение информации о торгах
      requestBody:
        description: данные для сортировки лотов
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sort:
                  $ref: '#/components/schemas/Sort'
      operationId: getTrades
      responses:
        "200":
          description: Информация о 20 последних торгах
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeInfo'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
  /trades/filter:
    put:
      parameters:
        - name: Content-Language
          in: header
          description: Заголовок для определения локали (ru, en, zh_CN)
          required: true
        - in: query
          name: page
          schema:
            type: integer
          required: true
          description: Номер страницы для серверной пагинации
      tags:
        - Trades
      summary: Получение информации о торгах в соответствии с фильтрами
      operationId: getFilteredTrades
      requestBody:
        description: Данные о выбранных фильтрах
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filters'
        required: true
      responses:
        "200":
          description: Информация о 20 или 5 лотах (в зависимости от того, автворизован пользователь или нет)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeInfo'
        "403":
          description: Неавторизованный пользователь просмотрел 5 лотов, выдаем сообщение о необходимости регистрации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
  /trades/nearest:
    put:
      parameters:
        - name: Content-Language
          in: header
          description: Заголовок для определения локали (ru, en, zh_CN)
          required: true
        - in: query
          name: page
          schema:
            type: integer
          required: true
          description: Номер страницы для серверной пагинации
      tags:
        - Trades
      summary: Получение информации о ближайших торгах в соответствии с фильтрами
      operationId: getNearestTrades
      requestBody:
        description: Данные о выбранных фильтрах
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filters'
        required: true
      responses:
        "200":
          description: Информация о 20 или 5 ближайших торгах (в зависимости от того, автворизован пользователь или нет)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeInfo'
        "403":
          description: Неавторизованный пользователь просмотрел 5 лотов, выдаем сообщение о необходимости регистрации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
  /trades/short/lots:
    put:
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          required: true
          description: Номер страницы для серверной пагинации
      tags:
        - Trades
      summary: Получение сокращенной информации о торгах по поиску
      requestBody:
        description: данные для поиска лотов
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                searchString:
                  type: string
                  example: лот№1
      operationId: getShortTrades
      responses:
        "200":
          description: Информация о 5 торгах по поиску
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TradeShortInfo'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
  /trades/victories:
    get:
      parameters:
        - name: Content-Language
          in: header
          description: Заголовок для определения локали (ru, en, zh_CN)
          required: true
        - in: query
          name: page
          schema:
            type: integer
          required: true
          description: Номер страницы для серверной пагинации
      tags:
        - Trades
      summary: Получение информации о победах недели
      operationId: getVictories
      responses:
        "200":
          description: Информация о 20 победах недели
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Victory'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
  /trades/{auctionId}:
    put:
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          required: true
          description: Номер страницы для серверной пагинации
        - name: Content-Language
          in: header
          description: Заголовок для определения локали (ru, en, zh_CN)
          required: true
        - in: path
          name: auctionId
          schema:
            type: integer
          required: true
          description: Айдишник торга, все лоты которого нужно получить
      tags:
        - Trades
      summary: Получение информации о лотах выбранного торга
      requestBody:
        description: данные для сортировки лотов
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sort:
                  $ref: '#/components/schemas/Sort'
      operationId: getLotsByAuction
      responses:
        "200":
          description: Информация о 20 лотах выбранного торга
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeInfo'
        "404":
          description: Торг по переданному айдишнику не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
...
  /trades/lot/{lotId}:
    get:
      parameters:
        - in: path
          name: lotId
          schema:
            type: integer
          required: true
          description: Айдишник лота, информацию о котором нужно получить
        - name: Content-Language
          in: header
          description: Заголовок для определения локали (ru, en, zh_CN)
          required: true
      tags:
        - Trades
      summary: Получение информации о лоте
      operationId: getLotInformation
      responses:
        "200":
          description: Информация о лоте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lot'
        "404":
          description: Лот по переданному айдишнику не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
  /trades/short/lot/{lotId}:
    get:
      parameters:
        - in: path
          name: lotId
          schema:
            type: integer
          required: true
          description: Айдишник лота, краткую информацию о котором нужно получить
      tags:
        - Trades
      summary: Получение краткой информации о лоте
      operationId: getShortLotInformation
      responses:
        "200":
          description: Информация о лоте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeShortInfo'
        "404":
          description: Лот по переданному айдишнику не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
  /trades/notifications/{lotId}:
    get:
      parameters:
        - in: path
          name: lotId
          schema:
            type: integer
          required: true
          description: Айдишник лота, сообщения о котором нужно получить
        - name: Content-Language
          in: header
          description: Заголовок для определения локали (ru, en, zh_CN)
          required: true
      tags:
        - Trades
      summary: Получение уведомлений о лоте
      operationId: getLotNotifications
      responses:
        "200":
          description: Информация об уведомлениях
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        "404":
          description: Лот по переданному айдишнику не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
...
  /trades/lot/toggle-seen/{lotId}:
    put:
      parameters:
        - in: path
          name: lotId
          schema:
            type: integer
          required: true
          description: Айдишник лота, который добавляем/удаляем в/из "просмотрено"
      tags:
        - Trades
      summary: Добавление/удаление лота в/из "просмотрено"
      operationId: toggleSeen
      responses:
        "200":
          description: Лот успешно добавлен/удален в/из "просмотрено"
        "404":
          description: Лот по переданному айдишнику не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
  /trades/lot/toggle-hide/{lotId}:
    put:
      parameters:
        - in: path
          name: lotId
          schema:
            type: integer
          required: true
          description: Айдишник лота, который добавляем/удаляем в/из "скрыто"
      tags:
        - Trades
      summary: Добавление/удаление лота в/из "скрыто"
      operationId: toggleHide
      responses:
        "200":
          description: Лот успешно добавлен/удален в/из "скрыто"
        "404":
          description: Лот по переданному айдишнику не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
  /trades/lot/toggle-pin/{lotId}:
    put:
      parameters:
        - in: path
          name: lotId
          schema:
            type: integer
          required: true
          description: Айдишник лота, который добавляем/удаляем в/из "закреплено"
      tags:
        - Trades
      summary: Добавление/удаление лота в/из "закреплено"
      operationId: togglePin
      responses:
        "200":
          description: Лот успешно добавлен/удален в/из "закреплено"
        "404":
          description: Лот по переданному айдишнику не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
  /trades/make/hidden/lots/{tradeId}:
    put:
      parameters:
        - in: path
          name: tradeId
          schema:
            type: integer
          required: true
          description: Айдишник торга, лоты которого нужно скрыть
      tags:
        - Trades
      summary: Скрытие всех связанных лотов
      operationId: makeHiddenLotByTrade
      responses:
        "200":
          description: Информация о лотах, которые был скрыты
          content:
            application/json:
              schema:
                type: object
                properties:
                  lotIds:
                    type: array
                    items:
                      type: integer
                      description: массив айдишников лотов, над которыми было совершено действие
        "404":
          description: Торги по переданному айдишнику не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
  /trades/delete/hidden/lots/{tradeId}:
    put:
      parameters:
        - in: path
          name: tradeId
          schema:
            type: integer
          required: true
          description: Айдишник торга, лоты которого нужно убрать из скрытого
      tags:
        - Trades
      summary: Удаление из скрытого всех связанных лотов
      operationId: deleteHiddenLotByTrade
      responses:
        "200":
          description: Информация о лотах, которые были удалены из скрытого
          content:
            application/json:
              schema:
                type: object
                properties:
                  lotIds:
                    type: array
                    items:
                      type: integer
                      description: массив айдишников лотов, над которыми было совершено действие
        "404":
          description: Торги по переданному айдишнику не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
...
components:
  schemas:
...
    TradeInfo:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/LotShort'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        count:
          type: integer
          description: количество выведенных записей
        perPage:
          type: integer
          description: количество записей на каждой странице
        currentPage:
          type: integer
          description: номер текущей страницы
        lastPage:
          type: integer
          description: номер последней страницы
        from:
          type: integer
          description: начало интервала
        to:
          type: integer
          description: конец интервала
        nextPageUrl:
          type: string
          description: ссылка на следующую страницу
        prevPageUrl:
          type: string
          description: ссылка на предыдущую страницу
...
    Error:
      required:
        - code
        - detail
        - title
      type: object
      properties:
        code:
          type: integer
          format: int32
        title:
          type: string
        detail:
          type: string
...
    Filters:
      type: object
      properties:
        sort:
          $ref: "#/components/schemas/Sort"
        categories:
          type: array
          default: null
          example: [ 'commercialRealEstate', 'land', 'obligations' ]
          items:
            type: string
            description: названия отмеченных категорий, если выбрана одна из главных категорий - присылать ее название и названия всех подкатегорий
        regions:
          type: array
          default: null
          example: [ 'MoscowCity', 'VoronezhRegion', 'KamchatkaKrai' ]
          items:
            type: string
            description: названия отмеченных регионов, если выбрана группа регионов - присылать все регионы этой группы
        prices:
          $ref: "#/components/schemas/Prices"
        dates:
          type: object
          default: null
          example: {
            eventTimeStart: {
              start: '23.06.2022',
              end: '23.08.2022'
            },
            eventTimeEnd: {
              start: '23.06.2022',
              end: '23.08.2022'
            },
            applicationTimeStart: {
              start: '23.06.2022',
              end: '23.08.2022'
            },
            applicationTimeEnd: {
              start: '23.06.2022',
              end: '23.08.2022'
            }
          }
          properties:
            eventTimeStart:
              type: object
              default: null
              properties:
                start:
                  type: string
                end:
                  type: string
            eventTimeEnd:
              type: object
              default: null
              properties:
                start:
                  type: string
                end:
                  type: string
            applicationTimeStart:
              type: object
              default: null
              properties:
                start:
                  type: string
                end:
                  type: string
            applicationTimeEnd:
              type: object
              default: null
              properties:
                start:
                  type: string
                end:
                  type: string
        extraOptions:
          type: object
          default: null
          example: {
            debtorCategories: [ 'SimplePerson', 'SimpleOrganization' ],
            debtors: [ 1, 2, 3 ],
            organizers: [ 1, 2, 3 ],
            arbitrationManagers: [ 1, 2, 3 ],
            other: { "period": "periodAll", "hasPhotos": false, "isHidden": false, "iStopped": false, "isCompleted": false }
          }
          properties:
            debtorCategories:
              type: array
              default: null
              items:
                type: string
                description: название категории должника
            debtors:
              type: array
              default: null
              example: [ 1, 2, 3 ]
              items:
                type: integer
                description: айдишники должников
            organizers:
              type: array
              default: null
              example: [ 1, 2, 3 ]
              items:
                type: integer
                description: айдишники организаторов торгов
            arbitrationManagers:
              type: array
              default: null
              example: [ 1, 2, 3 ]
              items:
                type: integer
                description: айдишники арбитражных управляющих
            other:
              type: object
              properties:
                period:
                  type: string
                  enum:
                    - periodAll
                    - periodDay
                    - periodWeek
                    - periodMonth
                  example: all
                hasPhotos:
                  type: boolean
                  example: false
                isStopped:
                  type: boolean
                  example: false
                isCompleted:
                  type: boolean
                  example: false
                isHidden:
                  type: boolean
                  example: false
        mainParams:
          type: object
          properties:
            excludedWords:
              type: string
              description: разделитель - запятая
              example: Машина, лот №1
            includedWords:
              type: string
              description: разделитель - запятая
              example: Машина, лот №1
            marks:
              type: array
              description: массив айдишников меток пользователя
            tradePlaces:
              type: array
              example: [ 1, 2, 3 ]
              description: массив айдишников торговых площадок
            tradeTypes:
              type: array
              items:
                type: string
                enum:
                  - OpenAuction
                  - OpenConcours
                  - PublicOffer
                  - CloseAuction
                  - CloseConcours
                  - ClosePublicOffer
                description: вид торгов
        searchString:
          type: string
          description: строка для поиска по лотам
...
    Images:
      type: array
      items:
        type: object
        properties:
          type:
            type: string
            enum:
              - system
              - user
          preview:
            type: string
          main:
            type: string
          id:
            type: integer
...
    Sort:
      type: object
      properties:
        direction:
          type: string
          enum:
            - asc
            - desc
        type:
          type: string
          enum:
            - publishDate
            - currentPrice
            - eventStart
            - eventEnd
            - applicationStart
            - applicationEnd
...
    Victory:
      type: object
      properties:
        tradeId:
          type: string
          description: номер торгов
        description:
          type: string
        photos:
          $ref: "#/components/schemas/Images"
        categories:
          $ref: "#/components/schemas/Categories"
        startPrice:
          type: integer
        endPrice:
          type: integer
        percentageReduction:
          type: integer
        percentageDirection:
          type: string
          enum:
            - up
            - down
          description: показатель того, значение percentageReduction является процентом падения иили процентом роста
        date:
          type: string
        participantsCount:
          type: integer
        winner:
          type: object
          properties:
            person:
              type: object
              properties:
                firstName:
                  type: string
                middleName:
                  type: string
                lastName:
                  type: string
            company:
              type: object
              properties:
                shortName:
                  type: string
                fullName:
                  type: string
            victoriesCount:
              type: integer
...
    TradeShortInfo:
      type: object
      properties:
        id:
          type: integer
          description: айди лота
        trade:
          $ref: '#/components/schemas/Trade'
        description:
          type: string
          description: описание лота
        lotNumber:
          type: string
          description: номер лота
        photos:
          $ref: "#/components/schemas/Images"
        categories:
          $ref: "#/components/schemas/Categories"
        currentPrice:
          type: number
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT